Update charm-manager for typescript
